backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pyyaml==6.0.1
python-multipart==0.0.6
jinja2==3.1.2
alembic==1.13.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
